requirements.txt
langchain 
langchain-community
langchain-openai
faiss-cpu
unstructured
unstructured[local-inference]
pandas
python-dotenv
requests
langchain-experimental
pymongo[srv]
langchain-mongodb
flask